Fire_Safety_Project/requirements.txt
opencv-python
numpy
matplotlib
paho-mqtt
pandas
scikit-learn
torch
torchvision
flask
